+++ pkg/commands/commands.go
@@ -78,6 +78,8 @@ import (
 	tlscustomprivatekey "github.com/fastly/cli/pkg/commands/tls/custom/privatekey"
 	tlsplatform "github.com/fastly/cli/pkg/commands/tls/platform"
 	tlssubscription "github.com/fastly/cli/pkg/commands/tls/subscription"
+	"github.com/fastly/cli/pkg/commands/tools"
+	domainTools "github.com/fastly/cli/pkg/commands/tools/domain"
 	"github.com/fastly/cli/pkg/commands/update"
 	"github.com/fastly/cli/pkg/commands/user"
 	"github.com/fastly/cli/pkg/commands/vcl"
@@ -486,6 +488,10 @@ func Define( // nolint:revive // function-length
 	tlsSubscriptionDescribe := tlssubscription.NewDescribeCommand(tlsSubscriptionCmdRoot.CmdClause, data)
 	tlsSubscriptionList := tlssubscription.NewListCommand(tlsSubscriptionCmdRoot.CmdClause, data)
 	tlsSubscriptionUpdate := tlssubscription.NewUpdateCommand(tlsSubscriptionCmdRoot.CmdClause, data)
+	toolsCmdRoot := tools.NewRootCommand(app, data)
+	toolsDomainCmdRoot := domainTools.NewRootCommand(toolsCmdRoot.CmdClause, data)
+	toolsDomainStatus := domainTools.NewDomainStatusCommand(toolsDomainCmdRoot.CmdClause, data)
+	toolsDomainSuggestions := domainTools.NewDomainSuggestionsCommand(toolsDomainCmdRoot.CmdClause, data)
 	updateRoot := update.NewRootCommand(app, data)
 	userCmdRoot := user.NewRootCommand(app, data)
 	userCreate := user.NewCreateCommand(userCmdRoot.CmdClause, data)
@@ -895,6 +901,10 @@ func Define( // nolint:revive // function-length
 		tlsSubscriptionDescribe,
 		tlsSubscriptionList,
 		tlsSubscriptionUpdate,
+		toolsCmdRoot,
+		toolsDomainCmdRoot,
+		toolsDomainStatus,
+		toolsDomainSuggestions,
 		updateRoot,
 		userCmdRoot,
 		userCreate,